requirements.txt
rembg==2.0.59
flask==3.1.0
pillow==11.1.0
gunicorn==23.0.0
